Implement random continent partitioning
Now the quantities of territories in each continent are randomly
chosen each run, and are guaranteed to be within certain bounds.

Now the internal structures of each continent need to be generated
and measured (in terms of diameter). Then the continents need to
be stitched together and measured again (in terms of macro-diameter)